Improve performance by parsing one declaration at a time, instead of the entire input at once
Since C++ grammar is highly ambiguous (esp. before the semantic actions can reject bad code), there is exponential explosion in the number of parses in the ASF forest if all statements are parse together. As a result, just parsing 20 lines was taking 60 seconds. By parsing each declaration separately, this now takes a fraction of a second.

Notice that the code parses declarations individually even if they are nested inside other declarations, such as

  namespace std {
      // lots-of nested declarations
  }